Move emoji rendering logic to util class
